+++ airbyte-integrations/bases/connector-acceptance-test/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.10.4
+Fixing bug: test_catalog_has_supported_data_types should support stream properties having `/` in it.
+
+## 0.10.3
+Fixing bug: test_catalog_has_supported_data_types , integer is a supported airbyte type.
+
 ## 0.10.2
 Fixing bug: test_catalog_has_supported_data_types was failing when a connector stream property is named 'type'.